// src/api/verifications/enums/guided-investigation.enum.ts
export enum GuidedQuestionType {
  SINGLE_SELECT = 'SINGLE_SELECT',
  MULTIPLE_SELECT = 'MULTIPLE_SELECT',
  SHORT_TEXT = 'SHORT_TEXT',
  PROVISIONAL_VERDICT = 'PROVISIONAL_VERDICT',
}

export enum GuidedQuestionCode {
  RESPONSIBLE_SHARING = 'RESPONSIBLE_SHARING',
  SOURCE_STRENGTH = 'SOURCE_STRENGTH',
  MISSING_CONTEXT = 'MISSING_CONTEXT',
  PROVISIONAL_VERDICT = 'PROVISIONAL_VERDICT',
}

export enum MediaLiteracyCompetency {
  CLAIM_IDENTIFICATION = 'CLAIM_IDENTIFICATION',
  SOURCE_IDENTIFICATION = 'SOURCE_IDENTIFICATION',
  SOURCE_INDEPENDENCE = 'SOURCE_INDEPENDENCE',
  DATE_AND_RECENCY = 'DATE_AND_RECENCY',
  EVIDENCE_EVALUATION = 'EVIDENCE_EVALUATION',
  CONTEXT_RECOGNITION = 'CONTEXT_RECOGNITION',
  MANIPULATION_RECOGNITION = 'MANIPULATION_RECOGNITION',
  VISUAL_VERIFICATION = 'VISUAL_VERIFICATION',
  AUDIO_VIDEO_CAUTION = 'AUDIO_VIDEO_CAUTION',
  RESPONSIBLE_SHARING = 'RESPONSIBLE_SHARING',
  AI_CONTENT_CAUTION = 'AI_CONTENT_CAUTION',
}

export enum GuidedInvestigationStatus {
  READY = 'READY',
  SUBMITTED = 'SUBMITTED',
  FEEDBACK_READY = 'FEEDBACK_READY',
}